+++ packages/atomic/src/components/search/result-lists/atomic-folded-result-list/atomic-folded-result-list.tsx
@@ -46,9 +46,7 @@ import {
   styleUrl: '../result-list-common.pcss',
   shadow: true,
 })
-export class AtomicFoldedResultList
-  implements BaseResultList<FoldedResult>, ResultListInfo
-{
+export class AtomicFoldedResultList implements BaseResultList, ResultListInfo {
   @InitializeBindings() public bindings!: Bindings;
   public foldedResultList!: FoldedResultList;
   public resultsPerPage!: ResultsPerPage;
@@ -72,7 +70,7 @@ export class AtomicFoldedResultList
   @FocusTarget() nextNewResultTarget!: FocusTargetController;
 
   public resultListCommon!: ResultListCommon;
-  private renderingFunction: ((res: FoldedResult) => HTMLElement) | null = null;
+  private renderingFunction: ResultRenderingFunction | null = null;
   private loadingFlag = randomID('firstResultLoaded-');
 
   /**
@@ -83,6 +81,7 @@ export class AtomicFoldedResultList
 
   /**
    * A list of non-default fields to include in the query results, separated by commas.
+   * @deprecated add it to atomic-search-interface instead
    */
   @Prop({reflect: true}) public fieldsToInclude = '';
   /**
@@ -120,9 +119,7 @@ export class AtomicFoldedResultList
    *
    * @param render
    */
-  @Method() public async setRenderFunction(
-    render: (result: FoldedResult) => HTMLElement
-  ) {
+  @Method() public async setRenderFunction(render: ResultRenderingFunction) {
     this.renderingFunction = render;
     this.assignRenderingFunctionIfPossible();
   }
@@ -162,7 +159,6 @@ export class AtomicFoldedResultList
     this.resultListCommon = new ResultListCommon({
       host: this.host,
       bindings: this.bindings,
-      fieldsToInclude: this.fieldsToInclude,
       templateElements: this.host.querySelectorAll('atomic-result-template'),
       onReady: () => {
         this.ready = true;
@@ -176,9 +172,14 @@ export class AtomicFoldedResultList
     });
 
     try {
-      this.foldedResultList = this.initFolding(
-        this.resultListCommon.resultListControllerProps
+      const localFieldsToInclude = this.fieldsToInclude
+        ? this.fieldsToInclude.split(',').map((field) => field.trim())
+        : [];
+      const fieldsToInclude = localFieldsToInclude.concat(
+        this.bindings.store.state.fieldsToInclude
       );
+
+      this.foldedResultList = this.initFolding({options: {fieldsToInclude}});
       this.resultsPerPage = buildResultsPerPage(this.bindings.engine);
     } catch (e) {
       this.error = e as Error;