+++ packages/atomic/src/components/search/result-lists/atomic-result-list/atomic-result-list.tsx
@@ -55,9 +55,7 @@ import {
   styleUrl: '../result-list-common.pcss',
   shadow: true,
 })
-export class AtomicResultList
-  implements BaseResultList<Result>, ResultListInfo
-{
+export class AtomicResultList implements BaseResultList, ResultListInfo {
   @InitializeBindings() public bindings!: Bindings;
   public resultList!: ResultList;
   public resultsPerPage!: ResultsPerPage;
@@ -87,11 +85,12 @@ export class AtomicResultList
   @FocusTarget() nextNewResultTarget!: FocusTargetController;
 
   public resultListCommon!: ResultListCommon;
-  private renderingFunction: ((res: Result) => HTMLElement) | null = null;
+  private renderingFunction: ResultRenderingFunction | null = null;
   private loadingFlag = randomID('firstResultLoaded-');
 
   /**
    * A list of non-default fields to include in the query results, separated by commas.
+   * @deprecated add it to atomic-search-interface instead
    */
   @Prop({reflect: true}) public fieldsToInclude = '';
   /**
@@ -122,9 +121,7 @@ export class AtomicResultList
    *
    * @param render
    */
-  @Method() public async setRenderFunction(
-    render: (result: Result) => HTMLElement
-  ) {
+  @Method() public async setRenderFunction(render: ResultRenderingFunction) {
     this.renderingFunction = render;
     this.assignRenderingFunctionIfPossible();
   }
@@ -165,7 +162,6 @@ export class AtomicResultList
     this.resultListCommon = new ResultListCommon({
       host: this.host,
       bindings: this.bindings,
-      fieldsToInclude: this.fieldsToInclude,
       templateElements: this.host.querySelectorAll('atomic-result-template'),
       onReady: () => {
         this.ready = true;
@@ -178,10 +174,16 @@ export class AtomicResultList
       nextNewResultTarget: this.nextNewResultTarget,
     });
 
-    this.resultList = buildResultList(
-      this.bindings.engine,
-      this.resultListCommon.resultListControllerProps
+    const localFieldsToInclude = this.fieldsToInclude
+      ? this.fieldsToInclude.split(',').map((field) => field.trim())
+      : [];
+    const fieldsToInclude = localFieldsToInclude.concat(
+      this.bindings.store.state.fieldsToInclude
     );
+
+    this.resultList = buildResultList(this.bindings.engine, {
+      options: {fieldsToInclude},
+    });
     this.resultsPerPage = buildResultsPerPage(this.bindings.engine);
     registerResultListToStore(this.bindings.store, this);
   }