+++ packages/headless/src/app/insight-engine/insight-engine.ts
@@ -10,7 +10,12 @@ import {
 } from '../engine';
 import {InsightThunkExtraArguments} from '../insight-thunk-extra-arguments';
 import {buildLogger} from '../logger';
-import {insightConfiguration, insightInterface, searchHub} from '../reducers';
+import {
+  insightConfiguration,
+  insightInterface,
+  search,
+  searchHub,
+} from '../reducers';
 import {buildThunkExtraArguments} from '../thunk-extra-arguments';
 import {
   InsightEngineConfiguration,
@@ -18,11 +23,16 @@ import {
 } from './insight-engine-configuration';
 import {Logger} from 'pino';
 import {setInsightConfiguration} from '../../features/insight-configuration/insight-configuration-actions';
+import {SearchAction} from '../../features/analytics/analytics-utils';
+import {logInterfaceLoad} from '../../features/analytics/analytics-actions';
+import {executeSearch} from '../../features/insight-search/insight-search-actions';
+import {firstSearchExecutedSelector} from '../../features/search/search-selectors';
 
 export type {InsightEngineConfiguration};
 
 const insightEngineReducers = {
   insightConfiguration,
+  search,
   insightInterface,
   searchHub,
 };
@@ -32,7 +42,14 @@ type InsightEngineState = StateFromReducersMapObject<InsightEngineReducers> &
   Partial<InsightAppState>;
 
 export interface InsightEngine<State extends object = {}>
-  extends CoreEngine<State & InsightEngineState, InsightThunkExtraArguments> {}
+  extends CoreEngine<State & InsightEngineState, InsightThunkExtraArguments> {
+  /**
+   * Executes the first search.
+   *
+   * @param analyticsEvent - The analytics event to log in association with the first search. If unspecified, `logInterfaceLoad` will be used.
+   */
+  executeFirstSearch(analyticsEvent?: SearchAction): void;
+}
 
 /**
  * The insight engine options.
@@ -87,6 +104,16 @@ export function buildInsightEngine(
 
     get state() {
       return engine.state;
+    },
+
+    executeFirstSearch(analyticsEvent = logInterfaceLoad()) {
+      const firstSearchExecuted = firstSearchExecutedSelector(engine.state);
+
+      if (firstSearchExecuted) {
+        return;
+      }
+
+      engine.dispatch(executeSearch(analyticsEvent));
     },
   };
 }