+++ packages/headless/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.74.1](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.74.0...@coveo/headless@1.74.1) (2022-06-16)
+
+**Note:** Version bump only for package @coveo/headless
+
+
+
+
+
+# [1.74.0](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.73.0...@coveo/headless@1.74.0) (2022-06-16)
+
+
+### Features
+
+* **atomic:** allow user to configure number of results in instant re… ([#2095](https://github.com/coveo/ui-kit/issues/2095)) ([e709c91](https://github.com/coveo/ui-kit/commit/e709c91745dc3d85d95d579f4624415a329d9cd3))
+
+
+
+
+
 # [1.73.0](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.72.1...@coveo/headless@1.73.0) (2022-06-13)